Integrated marching cubes in main scene
