--- Assets/Scripts/Solar System/PlanetBody.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0303f437bbf64174b997b70d614b32da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
